Add material UI submit button to theme (#1806)
