+++ packages/material-ui/src/Theme/Theme.tsx
@@ -1,0 +1,37 @@
+import React from "react";
+
+import Box from "@material-ui/core/Box";
+import Button from "@material-ui/core/Button";
+
+import ArrayFieldTemplate from "../ArrayFieldTemplate";
+import ErrorList from "../ErrorList";
+import Fields from "../Fields";
+import FieldTemplate from "../FieldTemplate";
+import ObjectFieldTemplate from "../ObjectFieldTemplate";
+import Widgets from "../Widgets";
+
+import { ThemeProps } from "@rjsf/core";
+import { utils } from "@rjsf/core";
+const { getDefaultRegistry } = utils;
+
+const { fields, widgets } = getDefaultRegistry();
+
+const DefaultChildren = () => (
+  <Box marginTop={3}>
+    <Button type="submit" variant="contained" color="primary">
+      Submit
+    </Button>
+  </Box>
+);
+
+const Theme: ThemeProps = {
+  children: <DefaultChildren />,
+  ArrayFieldTemplate,
+  fields: { ...fields, ...Fields },
+  FieldTemplate,
+  ObjectFieldTemplate,
+  widgets: { ...widgets, ...Widgets },
+  ErrorList,
+};
+
+export default Theme;
